+++ server/data/games.json
@@ -4,40 +4,41 @@
       "title": "Picolo App game 1",
       "size": 28,
       "elements": [
-      "Im Uhrzeigersinn: Jeder buschstabiert den Vornahmen der Person, die rechts von euch sitzt. 5 Schlücke für den/die Verlierer! {{this.name}}",
-      "Trink 2 Schlücke, wenn du schonmal Sex in der Öffentlichkeit hattest. Du Ekliger!",
-      "---Virus: ${name} muss ab jetzt mit französischem Akzent sprechen",
-      "${name}, postet ein Saufbild in seine/ihre Instagram Story und schreibt #picco oder trinkt sein/ihr Glas aus",
-      "Die nächste Person, die trinken muss, verteilt 4 Schlücke",
-      "Spiel: jeder wählt entweder Kopf oder Zahl, ${name} wirft die Münze. Alle die Kpf getippt habe trinken 4 Schlücke",
-      "${name}, leere 2 Gläser oder tinke 3 Schlücke",
-      "Trinke 3 Schlücke, wenn du dich diese Woche schon nackt im Spiegel angeschaut hast",
-      "Trinke 2 Schlücke wenn du schonmal einen Freund angepinkelt hast",
-      "${name} hast du schonmal Paintball gespielt? Wenn ja verteile 3 Schlücke, sonst trinke sie selber",
-      "Alle Männer mit langen Haaren und Frauen mit kurzen Haaren trinken ihr Glas auf Ex",
-      "${name}, verteile 4 Schlücke",
-      "---Niemand muss mehr mit Französischem Akzent reden",
-      "${name} verteile soviele Schlücke, wie du berühmte Physiker aufzählen kannst. Höre auf wann du willst, aber sagst du einen falschen Namen trinkst du soviele wie du verteilen wolltest",
-      "Spiel: Jeder der ein weises Kleidungsstück trägt trinkt 2 Schlücke",
-      "${name} schliese die Augen und sag, welche Farbe die Schuhe von ${name} haben, oder trinke 2 Schlucke",
-      "${name}, verteilt 3 Schlücke an eine Person, die seit start de Spiels schon pinkeln war",
-      "---Virus: beschinpft immer euren Nachbar, wenn ihr trinkt",
-      "${name}, wenn du ein Geschichte erzählen kannst, die darlegt, dass ${name} die größte Pfeife ist, verteile 4 Schlücke. Sonst trinke sie selber ---------------",
-      "Trinke 2 Schlücke, wenn deine Socken weiß sind",
-      "Alle die schonmal Slow getanzt haben, ohne Erfog beim bevorzugte Geschlecht zu habe, trinken 2 Schlücke",
-      "${name} macht 10 Liegestütz oder trnkt 2 Schlücke",
-      "---Hört auf euren Nachbar zu beschimpfen",
-      "Einmal im Kreis: Jeder sagt was er/sie schonmal gemacht hat. Alle die das nochnciht gemacht haben, triken. ${name} beginnt",
-      "Lieber Arm ab oder arm dran? Wählt gleichzeitig, die Verlierer trinken 2 Schlücke",
-      "${name} trinkt 2 Schlücke, wenn er letzten Monat Sex hatte",
-      "Alle Tabakraucher tinken 4 Schlücke" 
+        "${name} -- {{this.name}} -- {{name}} -- {{$t(name)}}",
+        "Im Uhrzeigersinn: Jeder buschstabiert den Vornahmen der Person, die rechts von euch sitzt. 5 Schlücke für den/die Verlierer! {{this.name}}",
+        "Trink 2 Schlücke, wenn du schonmal Sex in der Öffentlichkeit hattest. Du Ekliger!",
+        "---Virus: ${name} muss ab jetzt mit französischem Akzent sprechen",
+        "${name}, postet ein Saufbild in seine/ihre Instagram Story und schreibt #picco oder trinkt sein/ihr Glas aus",
+        "Die nächste Person, die trinken muss, verteilt 4 Schlücke",
+        "Spiel: jeder wählt entweder Kopf oder Zahl, ${name} wirft die Münze. Alle die Kpf getippt habe trinken 4 Schlücke",
+        "${name}, leere 2 Gläser oder tinke 3 Schlücke",
+        "Trinke 3 Schlücke, wenn du dich diese Woche schon nackt im Spiegel angeschaut hast",
+        "Trinke 2 Schlücke wenn du schonmal einen Freund angepinkelt hast",
+        "${name} hast du schonmal Paintball gespielt? Wenn ja verteile 3 Schlücke, sonst trinke sie selber",
+        "Alle Männer mit langen Haaren und Frauen mit kurzen Haaren trinken ihr Glas auf Ex",
+        "${name}, verteile 4 Schlücke",
+        "---Niemand muss mehr mit Französischem Akzent reden",
+        "${name} verteile soviele Schlücke, wie du berühmte Physiker aufzählen kannst. Höre auf wann du willst, aber sagst du einen falschen Namen trinkst du soviele wie du verteilen wolltest",
+        "Spiel: Jeder der ein weises Kleidungsstück trägt trinkt 2 Schlücke",
+        "${name} schliese die Augen und sag, welche Farbe die Schuhe von ${name} haben, oder trinke 2 Schlucke",
+        "${name}, verteilt 3 Schlücke an eine Person, die seit start de Spiels schon pinkeln war",
+        "---Virus: beschinpft immer euren Nachbar, wenn ihr trinkt",
+        "${name}, wenn du ein Geschichte erzählen kannst, die darlegt, dass ${name} die größte Pfeife ist, verteile 4 Schlücke. Sonst trinke sie selber ---------------",
+        "Trinke 2 Schlücke, wenn deine Socken weiß sind",
+        "Alle die schonmal Slow getanzt haben, ohne Erfog beim bevorzugte Geschlecht zu habe, trinken 2 Schlücke",
+        "${name} macht 10 Liegestütz oder trnkt 2 Schlücke",
+        "---Hört auf euren Nachbar zu beschimpfen",
+        "Einmal im Kreis: Jeder sagt was er/sie schonmal gemacht hat. Alle die das nochnciht gemacht haben, triken. ${name} beginnt",
+        "Lieber Arm ab oder arm dran? Wählt gleichzeitig, die Verlierer trinken 2 Schlücke",
+        "${name} trinkt 2 Schlücke, wenn er letzten Monat Sex hatte",
+        "Alle Tabakraucher tinken 4 Schlücke" 
       ]
   },
   {
-      "id": 1,
-      "title": "heloworld2",
-      "size": 10,
-      "elements": [
+    "id": 1,
+    "title": "heloworld2",
+    "size": 10,
+    "elements": [
       "21erste ${name} {{name}} {{this.name}} this.name ",
       "22zweite",
       "23jnovwunrovb",
@@ -48,6 +49,6 @@
       "28ownrc",
       "29or monre o",
       "200oiwroireir"
-      ]
+    ]
   }
 ]
